# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SkyReserve Flight Booking Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /airlines:
    get:
      operationId: listAirlines
      tags: [airlines]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Airline" }
    post:
      operationId: createAirline
      tags: [airlines]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AirlineInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Airline" }
  /airlines/{id}:
    get:
      operationId: getAirline
      tags: [airlines]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Airline" }
        "404": { description: Not found }
    patch:
      operationId: updateAirline
      tags: [airlines]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AirlineUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Airline" }
    delete:
      operationId: deleteAirline
      tags: [airlines]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
  /airlines/{id}/stats:
    get:
      operationId: getAirlineStats
      tags: [airlines]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AirlineStats" }
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      parameters:
        - { name: search, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Customer" }
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
    patch:
      operationId: updateCustomer
      tags: [customers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerAdminUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
    delete:
      operationId: deleteCustomer
      tags: [customers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
  /customers/{id}/bookings:
    get:
      operationId: getCustomerBookings
      tags: [customers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Booking" }
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
        - { name: all, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Booking" }
    post:
      operationId: createBooking
      tags: [bookings]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookingInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "409": { description: Seat already booked }
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
    patch:
      operationId: updateBooking
      tags: [bookings]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookingUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
  /bookings/{id}/cancel:
    post:
      operationId: cancelBooking
      tags: [bookings]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
  /bookings/{id}/ticket:
    get:
      operationId: getBookingTicket
      tags: [bookings]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
...
components:
  schemas:
...
    Airline:
      type: object
      required: [id, name, country]
      properties:
        id: { type: integer }
        name: { type: string }
        country: { type: string }
        logoUrl: { type: ["string", "null"] }
        createdAt: { type: string }
    AirlineInput:
      type: object
      required: [name, country]
      properties:
        name: { type: string, minLength: 1 }
        country: { type: string, minLength: 1 }
        logoUrl: { type: string }
    AirlineUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        country: { type: string, minLength: 1 }
        logoUrl: { type: string }
    AirlineStats:
      type: object
      required: [airlineId, fleetSize, activeFlights, totalBookings, totalRevenue]
      properties:
        airlineId: { type: integer }
        fleetSize: { type: integer }
        activeFlights: { type: integer }
        totalBookings: { type: integer }
        totalRevenue: { type: number }
    Flight:
      type: object
      required: [id, airlineId, airlineName, departure, arrival, departureTime, arrivalTime, status, aircraft, totalSeats, availableSeats, basePrice, durationMinutes]
      properties:
        id: { type: integer }
        airlineId: { type: integer }
        airlineName: { type: string }
        flightNumber: { type: string }
        departure: { type: string }
        arrival: { type: string }
        departureTime: { type: string }
        arrivalTime: { type: string }
        status: { type: string }
        aircraft: { type: string }
        totalSeats: { type: integer }
        availableSeats: { type: integer }
        basePrice: { type: number }
        durationMinutes: { type: integer }
...
    Customer:
      type: object
      required: [id, name, email, phone, role]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        role: { type: string }
        avatarUrl: { type: ["string", "null"] }
        darkModePref: { type: boolean }
        notifyEmail: { type: boolean }
        notifySms: { type: boolean }
        createdAt: { type: string }
...
    CustomerAdminUpdate:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        role: { type: string }
    Booking:
      type: object
      required: [id, customerId, customerName, flightId, bookingDate, seatNumber, cabinClass, status, price, bookingReference, flight]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        customerName: { type: string }
        flightId: { type: integer }
        bookingDate: { type: string }
        seatNumber: { type: string }
        cabinClass: { type: string }
        status: { type: string }
        price: { type: number }
        bookingReference: { type: string }
        passengerName: { type: string }
        passengerEmail: { type: string }
        passengerPhone: { type: string }
        flight: { $ref: "#/components/schemas/Flight" }
    BookingInput:
      type: object
      required: [flightId, seatNumber, cabinClass, passengerName, passengerEmail, passengerPhone]
      properties:
        flightId: { type: integer }
        seatNumber: { type: string, minLength: 1 }
        cabinClass: { type: string, minLength: 1 }
        passengerName: { type: string, minLength: 1 }
        passengerEmail: { type: string, minLength: 1 }
        passengerPhone: { type: string, minLength: 1 }
        paymentMethodId: { type: integer }
    BookingUpdate:
      type: object
      properties:
        seatNumber: { type: string }
        status: { type: string }
        cabinClass: { type: string }
    Ticket:
      type: object
      required: [bookingId, bookingReference, passengerName, flight, seatNumber, cabinClass, qrPayload, status]
      properties:
        bookingId: { type: integer }
        bookingReference: { type: string }
        passengerName: { type: string }
        flight: { $ref: "#/components/schemas/Flight" }
        seatNumber: { type: string }
        cabinClass: { type: string }
        qrPayload: { type: string }
        status: { type: string }
...