# lib/api-zod/src/generated/types/airlineInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * SkyReserve Flight Booking Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface AirlineInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  country: string;
  logoUrl?: string;
}