# lib/api-zod/src/generated/types/bookingUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * SkyReserve Flight Booking Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingUpdate {
  seatNumber?: string;
  status?: string;
  cabinClass?: string;
}